+++ src/packages/refresh/index.ts
@@ -1,0 +1,6 @@
+/**
+ * 刷新
+ */
+export const refresh = (): void => {
+  window.location.reload()
+}