+++ src/packages/start-move-end/index.ts
@@ -1,0 +1,8 @@
+/**
+ * 把数组的第一项放到最后一项
+ * @param arr 
+ * @returns 
+ */
+export const startMoveEnd = (arr: []): [] => {
+  return arr.unshift(arr.pop())
+}