+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 更新日志
 
+- 新增 `start-move-end` `refresh` `end-move-start`
+
 ## 0.0.1 (2022-09-30)
 
 - 更新配置文件